Wall: Remove sea barrier model.
This model is awesome & is only being removed due to a lack of engine
support. The area build issue can be worked around by forcing this model
on the land. But there's currently no way offset collision box so
that'll cover both the surface and underwater. So for now we'll have to
stick with just the land barrier as it visually represents the actual
collision box.