Mine: Area build in strips instead of circles.
Mikey wont fix/allow it so client mods can be loaded after server mods.
So we're forced to include this within the balance.

    *   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
      *   *   *   *   *   *   *   *   *   *   *   *   *   *   *
    *   *   *   *   *   *   *   *   *   *   *   *   *   *   *   *

I went for a three line spaced pattern to best accommodate people that
people that want the old coverage while still providing the strategic
control players that prefer strips are accustomed to.